Commit sửa lại phần update Tower
Commit sửa lại phần update Tower

+++ Assets/Scripts/GamePlay/Level/UnitInfo.cs
@@ -11,7 +11,7 @@ public class UnitInfo : MonoBehaviour
     public string secondaryText;
 
     [HideInInspector]
-    public decimal level = 1;
+    public int level = 1;
 
 
     //For Tower
